Incompatibility checkers refactoring
