+++ sdk/src/main/java/com/horizen/transaction/FeePaymentsTransaction.java
@@ -4,6 +4,7 @@ import com.horizen.box.BoxUnlocker;
 import com.horizen.proposition.PublicKey25519Proposition;
 import com.horizen.box.ZenBox;
 import com.horizen.transaction.exception.TransactionSemanticValidityException;
+import com.horizen.transaction.incompatibilitychecker.MempoolIncompatibleTransactionIncompatibilityChecker;
 import scala.Array;
 import sparkz.core.serialization.SparkzSerializer;
