+++ sdk/src/main/java/com/horizen/transaction/OpenStakeTransaction.java
@@ -10,6 +10,7 @@ import com.horizen.proposition.PublicKey25519Proposition;
 import com.horizen.secret.PrivateKey25519;
 import com.horizen.secret.Secret;
 import com.horizen.transaction.exception.TransactionSemanticValidityException;
+import com.horizen.transaction.incompatibilitychecker.OpenStakeTransactionIncompatibilityChecker;
 import com.horizen.utils.Pair;
 import scala.Array;
