+++ sdk/src/main/java/com/horizen/transaction/incompatibilitychecker/DefaultTransactionIncompatibilityChecker.java
@@ -1,6 +1,8 @@
-package com.horizen.transaction;
+package com.horizen.transaction.incompatibilitychecker;
 
 import com.horizen.box.BoxUnlocker;
+import com.horizen.transaction.BoxTransaction;
+import com.horizen.transaction.TransactionIncompatibilityChecker;
 import com.horizen.utils.ByteArrayWrapper;
 
 import java.util.List;
