+++ sdk/src/main/java/com/horizen/transaction/incompatibilitychecker/OpenStakeTransactionIncompatibilityChecker.java
@@ -1,8 +1,12 @@
-package com.horizen.transaction;
+package com.horizen.transaction.incompatibilitychecker;
+
+import com.horizen.transaction.BoxTransaction;
+import com.horizen.transaction.OpenStakeTransaction;
+import com.horizen.transaction.incompatibilitychecker.DefaultTransactionIncompatibilityChecker;
 
 import java.util.List;
 
-public class OpenStakeTransactionIncompatibilityChecker extends DefaultTransactionIncompatibilityChecker{
+public class OpenStakeTransactionIncompatibilityChecker extends DefaultTransactionIncompatibilityChecker {
     @Override
     public <T extends BoxTransaction> boolean isTransactionCompatible(T newTx, List<T> currentTxs) {
         if (!super.isTransactionCompatible(newTx, currentTxs)) {
